+++ docs/Changelog.md
@@ -4,7 +4,10 @@ This page is a brief overview of each version.
 
 ---
 ## Unreleased
-- Nothing yet
+- Don't index and generate embeddings for pages in Library/
+- Add new `AI: Enhance Note` command to call existing `AI: Tag Note` and `AI: Suggest Page Name` commands on a note, and the new frontmatter command
+- Add new `AI: Generate Note FrontMatter` command to extract useful facts from a note and add them to the frontmatter
+- Always include the note’s current name in [[Commands/AI: Suggest Page Name]] as an option.
 
 ---
 ## 0.2.0